Make the delete target EXTREMELY large and tweak tapping algorithm
Experimenting with a really large tap target for the delete icon.
Also tweaked the search algorithm to start at the beginning of
the spans when looking for a chip.
Also was able to get rid of the handler and delayed focus since
the code now watches for selection changes and always keeps
the cursor just past the last chip.
Change-Id: I105a7cb72ceb77f4f21b662a978e819fba3d9875