Added invalid chips, fixed comments.
Add an invalid background for chips that cannot be validated.
If there is a validator, and a chip is invalid, set it to a red
background for now.
Note: I left the old method for setChipDimensions so that
I don't break Gmail.
Updated comments in the code to be clearer.
Change-Id: I01553cf0c3b3a6c7c59da5016d32be364535f7f2