Fix the offset of the popup window.
If I use the view as the anchor and the vertical offset, then
the popup appears correctly even after a scroll.
I believe this will also take care of the cases where the popup
had been appearing above chips, as the view will be scrolled
to fit the popup if possible.

Fixes Bug:4723677 chips: sometimes popup appears ABOVE the chips field
Change-Id: I976be8e4e06ff44a093f8a93c65a63d433b1e774