Enable pasting of text into the recipient edit text field.
The issue was that we were hiding the cursor whenever the field was touched
and it should have just been when the field was touched AND a chip was selected.

fixes Bug:5016914 ICS email: cannot paste text into "to:" field in portrait mode compose message

Also fix the case where there is content and the user changes focus.
This had not been committing as committing chips depended on having focus.
Change-Id: I43fae67a399ae7df4862861144ed42b608718a2c