Area code truncated for the phone number in messaging app
Bug 6717762

When launching mms from the contacts app, numbers sometimes get munged,
either by losing the area code (if in parens) or losing the number
completely and just ending up with a contact name.
Parse all the parts of a contact "fred flinstone <512-123-1231>" looking
for the phone number. Also handle the case where the area code is lost when
in parens (i.e. (800) 891-8923).

Change-Id: I76976fa8004c1d7c03a651b3915537245fee16e4