Don't jump focus out of the View when the Activity starts
In an Activity where a RecipientEditTextView has initial focus, if
you start the Activity by pressing the enter key on a keyboard (like
on a button), it would appear the onKeyUp event is handled in the new
Activity, causing this View's enter-handling logic to run.

By moving the logic into onKeyDown, we can avoid this situation.

Bug: 8013052
Change-Id: I48d3e903071954f5b73f4860ea87733c6eacd2dd