Allow support for alternate data sources.
BaseRecipientAdapter adds a bunch of methods that can be
overridden to change the default behavior. Also makes a bunch
of private methods protected so that derived classes can use those
methods.

Abstracts the calls to update the various internal data structures
so that derived classes can utilize their own structures by just
overriding a few methods.

Refactors photo loading into the PhotoManager interface
with a default implementation living in DefaultPhotoManager.

Provides the ability to not use the RecipientAlternatesAdapter
and instead just use SingleRecipientAddressAdapter when showing
a dropdown after clicking on a chip.

Also allows overriding the default behavior for creating chips
when pasting in email addresses or editing drafts.

b/11387812

Change-Id: I826c30edeee1752b7a3d4108e777f90b79d0f7f8