Move getAvatarIcon off UI
Change BaseRecipientAdapter#fetchPhoto to use the async version of
populatePhotoBytes. Move some UI code around so we keep a reference
of each chip bitmap as well as criterias we need to draw the icon
once it's loaded. Once the icon bitmap is asynchronously loaded,
we then draw that icon onto the UI.

b/15082721

Change-Id: I8abd3f8e1bb61b9a5b35ecda71704bb0c38dc7bf