Chips show icon even if name is null
Not sure why this was added or why it was working before.
A recipient's name can be null but it can have a valid
destination, in which case we still want to show the icon.

In any case, based on the comments it says we only not
show the icon for generated/invalid contacts, so I'm
removing the check for empty name.

b/16383133

Change-Id: I16642986e503a51e52275118bea05f34973817b6