Bug 25022571: Disable long-press in accessibility mode
When the accessibility focus is on the chips text view and you do a
long-press, it shows the chip selection popup for the last chip. Selecting
the last chip is arbitrary (and there is no way to select a specific chip -
see Bug 25021253). Also, the normal gesture to select a chip is a tap,
not a long-press. Hence we disable the selection popup completely.

Change-Id: Ia5a890edafdbb01cb6a578aa009e7f4d82f80706