Add ability to show warning icon with warning text in chips.
The caller of RecipientEditTextView can specify a set of addresses to show
a warning icon for. It must also provide a warning icon bitmap and some warning
text. That icon will appear in the chip, and when the user touches it, the text
will show.

For more context on this change, see the bug.

Bug: 34170077
Test: manual - built in google3 and tried on my device.
Change-Id: I0067b5e9176c11a610dfe5580b415569de32d814